goby message chunk parsing.
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@23879 60cd2ccc-2b05-0410-9566-da273a940bff

+++ cpp/src/goby/Alignments.cc
@@ -94,7 +94,7 @@ namespace goby {
 
     // the stream may not get all read in at once so we may need to copy in chunks
     void* header = NULL;
-    int headerSize = 0;
+    size_t headerSize = 0;
 
     const void* buffer;
     int bufferSize;
